+++ lib/types.ts
@@ -35,3 +35,42 @@ export interface IterablePredicateCallback<T> {
 export interface IterableCircular<T> extends Iterable<T> {
   [Symbol.iterator](): IterableIterator<T>;
 }
+
+/**
+ * Type for all iterable methods in the library.
+ * @typeParam T - The type of items in the iterable argument.
+ * @typeParam U - The return type of the method.
+ * @typeParam Args - The rest type of other method arguments.
+ */
+export interface IterableMethod<T, U, Args extends unknown[]> {
+  (it: Iterable<T>, ...args: Args): U;
+}
+
+/**
+ * Alias for an `IterableMethod` which accepts no arguments accept an iterable.
+ * @typeParam T - The type of items in the iterable argument.
+ * @typeParam U - The return type of the method.
+ */
+export type UniaryIterableMethod<T, U> = IterableMethod<T, U, []>;
+
+/**
+ * A curried version of an `IterableMethod`. All methods in `fp.ts` are either
+ * of type `CurriedIterableMethod` or `UniaryIterableMethod`.
+ * @typeParam T - The type of items in the iterable argument.
+ * @typeParam U - The return type of the method.
+ * @typeParam Args - The rest type of other method arguments.
+ */
+export interface CurriedIterableMethod<T, U, Args extends unknown[]> {
+  (...args: Args): UniaryIterableMethod<T, U>;
+}
+
+/**
+ * An alias for a special case of `UniaryIterableMethod` for when the result is
+ * an iterable.
+ * @typeParam T - The type of items in the iterable argument.
+ * @typeParam U - The type of items in the returned iterable.
+ */
+export type UniaryIterableTransformer<T, U> = UniaryIterableMethod<
+  T,
+  Iterable<U>
+>;